PDP11: Reworked RH to device mapping for proper static address assignment
Operating systems and boot driver code presume that each Massbus
device will always be at a specific fixed address and use a specific fixed vector.
The prior implementation assigned the first active Massbus device to RHA
and followed in order for subsequent devices.  This could leave devices at
addresses and vectors that software or boot drivers wouldn't expect.

The problem which was originally reported in #301 should now be properly
solved.  As a consequence of the need to have fixed addresses and vectors,
some combinations of devices can not coexist in a system without causing
conflicts with each other.  TM and TU have conflicting interrupts.  TM and
TS have conflicting addresses and interrupts.